# Conflicts:
#	.idea/caches/gradle_models.ser
#	app/src/main/java/com/example/jvmori/myweatherapp/data/current/response/Clouds.java
#	app/src/main/java/com/example/jvmori/myweatherapp/data/current/response/Coord.java
#	app/src/main/java/com/example/jvmori/myweatherapp/data/current/response/CurrentWeatherResponse.java
#	app/src/main/java/com/example/jvmori/myweatherapp/data/current/response/Sys.java
#	app/src/main/java/com/example/jvmori/myweatherapp/data/current/response/Wind.java

+++ app/src/main/java/com/example/jvmori/myweatherapp/data/current/response/Wind.java
@@ -3,18 +3,19 @@ package com.example.jvmori.myweatherapp.data.current.response;
 
 import com.google.gson.annotations.SerializedName;
 
+
 public class Wind {
 
     @SerializedName("deg")
-    private Long mDeg;
+    private Double mDeg;
     @SerializedName("speed")
     private Double mSpeed;
 
-    public Long getDeg() {
+    public Double getDeg() {
         return mDeg;
     }
 
-    public void setDeg(Long deg) {
+    public void setDeg(Double deg) {
         mDeg = deg;
     }
